feat: make account creation atomic with INITIALIZING status and FAILED compensation
