+++ src/modules/accounts/enums/account-status.enum.ts
@@ -1,0 +1,8 @@
+export enum AccountStatus {
+  INITIALIZING = 'initializing',
+  PENDING_PAYMENT = 'pending_payment',
+  PENDING_CLAIM = 'pending_claim',
+  CLAIMED = 'claimed',
+  EXPIRED = 'expired',
+  FAILED = 'failed',
+}
