+++ src/modules/claims/providers/token-verification.provider.ts
@@ -12,8 +12,8 @@ import jwt from 'jsonwebtoken';
 import { TokenExpiredError, JsonWebTokenError } from 'jsonwebtoken';
 import * as crypto from 'crypto';
 import { Account } from '../../accounts/entities/account.entity.js';
-import { AccountStatus } from '../../accounts/entities/account.entity.js';
 import { ClaimVerificationResponseDto } from '../dto/claim-verification-response.dto.js';
+import { AccountStatus } from '@/modules/accounts/enums/account-status.enum.js';
 
 interface ClaimTokenPayload {
   publicKey: string;
