+++ src/modules/stellar/providers/payment-monitor-provider.ts
@@ -4,10 +4,8 @@ import { Repository } from 'typeorm';
 import { ConfigService } from '@nestjs/config';
 import * as StellarSdk from '@stellar/stellar-sdk';
 import { StellarService } from '../stellar.service.js';
-import {
-  Account,
-  AccountStatus,
-} from '../../../modules/accounts/entities/account.entity.js';
+import { Account } from '../../../modules/accounts/entities/account.entity.js';
+import { AccountStatus } from '@/modules/accounts/enums/account-status.enum.js';
 
 /**
  * PaymentMonitorProvider - Horizon SSE-based payment detection
